requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
